+++ app/Http/Controllers/CategoriesController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Category;
 use Session;
+use App\Category;
+use App\Product;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class CategoriesController extends Controller
@@ -15,6 +17,7 @@ class CategoriesController extends Controller
      */
     public function index()
     {
+        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
         $categories  = Category::get();
         return view("admin.categories.index")->with("categories", $categories);
     }
@@ -25,7 +28,8 @@ class CategoriesController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
+    {   
+        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
         $mode = ["edit" => false, "add" => true];
         return view("admin.categories.form")->with("mode", $mode);
     }
@@ -38,6 +42,7 @@ class CategoriesController extends Controller
      */
     public function store(Request $request, Category $category)
     {
+        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
         $data = [];
         $data["name"] = $request->input("name");
         if($request->isMethod("post")) {
@@ -61,8 +66,8 @@ class CategoriesController extends Controller
      */
     public function show($id)
     {
-        
-
+        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+        return redirect("/admin/categories");
     }
 
     /**
@@ -73,6 +78,7 @@ class CategoriesController extends Controller
      */
     public function edit($id)
     {
+        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
         $category = Category::find($id);
         $mode = ["edit" => true, "add" => false];
         return view("admin.categories.form")->with(["mode" => $mode, "category" => $category]);
@@ -87,6 +93,7 @@ class CategoriesController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
         $category = Category::find($id);
         if($request->isMethod("PUT")) 
         {
@@ -111,7 +118,20 @@ class CategoriesController extends Controller
      */
     public function destroy($id)
     {
+        if(Auth::user()->role !== 'admin') { return redirect("/admin"); }
+        $products = Product::get();
         $category = Category::find($id);
+        // foreach($products as $product) {
+        //     if($product->category_id === $category->id) {
+        //         foreach($product->images as $image) {
+        //             $file = public_path("images/products/").$image;
+        //             if(file_exists($file)) {
+        //                 unlink($file);
+        //             }
+        //         }
+        //         $product->delete();
+        //     }
+        // }
         $category->delete();
         Session::flash("success", "Category was deleted successfully");
         return redirect("/admin/categories");
